+++ Gruntfile.js
@@ -18,6 +18,7 @@ module.exports = function (grunt) {
     grunt.loadNpmTasks('grunt-contrib-less');
     grunt.loadNpmTasks('grunt-contrib-uglify');
     grunt.loadNpmTasks('grunt-contrib-watch');
+    grunt.loadNpmTasks('grunt-exec');
 
     var configBridge = grunt.file.readJSON('./bower_components/bootstrap/grunt/configBridge.json', {
         encoding: 'utf8'
@@ -161,6 +162,9 @@ module.exports = function (grunt) {
                     base: '.'
                 }
             }
+        },
+        exec: {
+            publish_public: 'php ../../../artisan vendor:publish --provider="Redooor\\Redminportal\\RedminportalServiceProvider" --tag="public" --force'
         }
     });
 
@@ -172,6 +176,9 @@ module.exports = function (grunt) {
     // Distribute all assets to public folder
     grunt.registerTask('dist-assets', ['copy:fonts', 'copy:jquery', 'copy:bootstrapjs', 'copy:bootstrapcss', 'copy:jqueryui', 'copy:jqueryuijs', 'copy:momentjs', 'copy:datetimepickerjs', 'copy:datetimepickercss', 'copy:redmaterialsjs', 'copy:redmaterialscss', 'copy:fontawesomefonts', 'copy:fontawesomecss']);
     
+    // Publish to public
+    grunt.registerTask('publish-assets', ['exec:publish_public']);
+    
     // Default task, compile and distribute all assets to public folder
-    grunt.registerTask('default', ['less-compile', 'dist-assets']);
+    grunt.registerTask('default', ['less-compile', 'dist-assets', 'publish-assets']);
 };