+++ CHANGELOG.md
@@ -28,6 +28,8 @@ This update introduces some new tables. Refer to [UPGRADE.md](UPGRADE.md) for th
 1. Added product variations. You can now add variations to each product. (issue #125)
 2. Added reusable partial templates for Modal Window and Language selector tab. (issue #130)
 3. Added shipping properties to products: Weight, Length, Width, Height, units. (issue #126)
+4. Fixed UI: Category hierarchy list word wrap. (issue #133)
+5. Developer: use grunt to publish assets to public folder automatically. (issue #139)
 
 ### Enhancements: